refactor the file-upload component to different parts
