+++ src/components/file-upload/index.ts
@@ -1,0 +1,6 @@
+export { FileUpload } from "./file-upload";
+export { ModelUpload } from "./model-upload";
+export { useFileUpload } from "./use-file-upload";
+export type { UseFileUploadOptions, UseFileUploadReturn } from "./use-file-upload";
+export type { UploadedFile, UploadingFile, FileUploadProps, ModelUploadProps } from "./types";
+export { getFileIcon, formatFileSize } from "./utils";
