+++ src/components/file-upload/utils.ts
@@ -1,0 +1,36 @@
+import {
+  File,
+  FileText,
+  Image as ImageIcon,
+  Video,
+  Music,
+  Archive,
+} from "lucide-react";
+
+export const getFileIcon = (type: string) => {
+  if (type.startsWith("image/")) return ImageIcon;
+  if (type.startsWith("video/")) return Video;
+  if (type.startsWith("audio/")) return Music;
+  if (
+    type.includes("zip") ||
+    type.includes("rar") ||
+    type.includes("7z") ||
+    type.includes("tar")
+  )
+    return Archive;
+  if (
+    type.includes("pdf") ||
+    type.includes("document") ||
+    type.includes("text")
+  )
+    return FileText;
+  return File;
+};
+
+export const formatFileSize = (bytes: number) => {
+  if (bytes === 0) return "0 Bytes";
+  const k = 1024;
+  const sizes = ["Bytes", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return Math.round((bytes / Math.pow(k, i)) * 100) / 100 + " " + sizes[i];
+};
